requirements.txt
# LLM APIクライアント
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.18.0

# ローカルLLM（オプション）
ollama>=0.1.0

# 共通ライブラリ
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
langchain>=0.1.0

# 論文検索
arxiv>=2.1.0
scholarly>=1.7.0

# その他
python-dotenv>=1.0.0

